refactor(online-eval): move criteria resolution out of the span producer
resolve_criteria and resolve_criteria_bulk are the DB-backed half of the
fingerprint recipe: the session sweeper, the executor's staleness guard, and the
span producer all resolve through them, yet they lived inside the span producer,
so two of the three callers imported a peer daemon to compute a shared value.
derivation's own docstring cedes version resolution and DB access to "callers",
which had come to mean the span producer specifically.

They move to criteria_resolution.py alongside _sandbox_runtime_fingerprint and
_resolved_criteria, together with the materialization stamp both materializers
write to the same criteria row. Import-only churn otherwise.
